+++ web/src/app/assistants/SidebarWrapper.tsx
@@ -10,7 +10,7 @@ import FixedLogo from "../../components/logo/FixedLogo";
 import { SettingsContext } from "@/components/settings/SettingsProvider";
 import { useChatContext } from "@/components/context/ChatContext";
 import { HistorySidebar } from "@/app/chat/components/sessionSidebar/HistorySidebar";
-import { useAssistants } from "@/components/context/AssistantsContext";
+import { useAssistantsContext } from "@/components/context/AssistantsContext";
 import AssistantModal from "./mine/AssistantModal";
 import { useSidebarShortcut } from "@/lib/browserUtilities";
 import { UserSettingsModal } from "@/app/chat/components/modal/UserSettingsModal";
@@ -45,7 +45,7 @@ export default function SidebarWrapper<T extends object>({
 
   const sidebarElementRef = useRef<HTMLDivElement>(null);
   const { folders, openedFolders, chatSessions } = useChatContext();
-  const { assistants } = useAssistants();
+  const { assistants } = useAssistantsContext();
   const explicitlyUntoggle = () => {
     setShowDocSidebar(false);
