+++ web/src/app/assistants/mine/AssistantSharingPopover.tsx
@@ -14,7 +14,7 @@ import { usePopup } from "@/components/admin/connectors/Popup";
 import { Bubble } from "@/components/Bubble";
 import { AssistantIcon } from "@/components/assistants/AssistantIcon";
 import { Spinner } from "@/components/Spinner";
-import { useAssistants } from "@/components/context/AssistantsContext";
+import { useAssistantsContext } from "@/components/context/AssistantsContext";
 
 interface AssistantSharingPopoverProps {
   assistant: Persona;
@@ -29,7 +29,7 @@ export function AssistantSharingPopover({
   allUsers,
   onClose,
 }: AssistantSharingPopoverProps) {
-  const { refreshAssistants } = useAssistants();
+  const { refreshAssistants } = useAssistantsContext();
   const { popup, setPopup } = usePopup();
   const [isUpdating, setIsUpdating] = useState(false);
   const [selectedUsers, setSelectedUsers] = useState<MinimalUserSnapshot[]>([]);
