+++ web/src/app/chat/components/sessionSidebar/HistorySidebar.tsx
@@ -29,7 +29,7 @@ import { pageType } from "./types";
 import LogoWithText from "@/components/header/LogoWithText";
 import { Persona } from "@/app/admin/assistants/interfaces";
 import { DragEndEvent } from "@dnd-kit/core";
-import { useAssistants } from "@/components/context/AssistantsContext";
+import { useAssistantsContext } from "@/components/context/AssistantsContext";
 import { AssistantIcon } from "@/components/assistants/AssistantIcon";
 import { buildChatUrl } from "@/app/chat/services/lib";
 import { reorderPinnedAssistants } from "@/lib/assistants/updateAssistantPreferences";
@@ -193,7 +193,7 @@ export const HistorySidebar = forwardRef<HTMLDivElement, HistorySidebarProps>(
     const router = useRouter();
     const { user, toggleAssistantPinnedStatus } = useUser();
     const { refreshAssistants, pinnedAssistants, setPinnedAssistants } =
-      useAssistants();
+      useAssistantsContext();
 
     const currentChatId = currentChatSession?.id;
