+++ web/src/app/ee/assistants/stats/[id]/AssistantStats.tsx
@@ -6,7 +6,7 @@ import {
   DateRangeSelector,
   DateRange,
 } from "@/app/ee/admin/performance/DateRangeSelector";
-import { useAssistants } from "@/components/context/AssistantsContext";
+import { useAssistantsContext } from "@/components/context/AssistantsContext";
 import { AssistantIcon } from "@/components/assistants/AssistantIcon";
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
 import { AreaChartDisplay } from "@/components/ui/areaChart";
@@ -26,7 +26,7 @@ type AssistantStatsResponse = {
 export function AssistantStats({ assistantId }: { assistantId: number }) {
   const [assistantStats, setAssistantStats] =
     useState<AssistantStatsResponse | null>(null);
-  const { assistants } = useAssistants();
+  const { assistants } = useAssistantsContext();
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [dateRange, setDateRange] = useState<DateRange>({
